move schemas to their own file
